Solve the problem with upstreams
Upstreams is no longer allowed if you have alphaConfig enabled, because
the option is available in alphaConfig. However, setting the default in
alphaConfig won't work, because people may have it disabled, and it's
disabled by default. This means we need this value to remain in
configFile as a default if alphaConfig isn't enabled.

Fixes #226